feat(TextField): add type password and resech
- add type password
- update react-testarea-autosize
- remove native required

closes #2351 #2350 #2354
