+++ src/components/TextField/TextField.css
@@ -346,18 +346,33 @@
     transition: background-color 0.15s;
   }
 
-  &-ClearButton {
+  &-ClearButton,
+  &-EyeButton {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
     height: calc(var(--input-height) - (var(--control-border-width) * 2));
     margin: 0;
     margin-left: var(--input-space);
     padding: 0;
-    color: var(--clear-button-color);
     background-color: transparent;
     border: none;
     cursor: pointer;
+  }
+
+  &-ClearButton {
+    color: var(--clear-button-color);
 
     &:hover {
       --clear-button-color: var(--color-control-typo-clear-hover);
+    }
+  }
+
+  &-EyeButton {
+    color: var(--color-control-typo-placeholder);
+
+    &:hover {
+      color: var(--color-control-typo-clear-hover);
     }
   }
